from-zero-to-hero/requirements-rc3.txt
agent-framework==1.0.0rc3
azure-ai-agentserver-agentframework==1.0.0b17
azure-monitor-opentelemetry